I don't think we should be calling 'toJava' on regular "in" arguments.
They are pointless at best, because those arrays come from java.lang.proxy code that packs interface invocations into an rray and so no one will look at the values afterward.

In some other cases, they are actively harmful. For example, NativeType.ComObject.toJava creates a wrapper object (which is the right thing to do for the original purpose of this method, which is to massage the 'out' parameter. When applied to the 'in' parameter this creates garbages that hold onto the reference until GCed.

Also see https://github.com/kohsuke/com4j/issues/4