Avoid new Variant for each index during iteration
Since a Variant wraps a DirectByteBuffer which can be costly,
avoid creating a new variant for each fetch() operation, and
instead use one instance that has its contents replaced for each
fetch().

This should be safe since next() extracts the contents of the
Variant before returning, creating a new reference to the contents
of the Variant.