Clean up duplicate refs in ComObjectCollector
ComObjectCollector stored its objects to collect in a WeakHashMap.
This caused problems in the case that multiple Wrapper instances
wrapped the same IUnknown*.  In that case, the objects would
hashCode() and equals() the same, and hence only one of the
Wrapper's would have dispose() called.  This would lead to COM
objects not being collected until the GC ran and reclaimed the
duplicate Wrapper instances.

Changed to store disposee's in a LinkedList of WeakReferences.  There
is no need to enforce uniqueness since every object that is passed to
onNewObject wraps a distinct copy of the pointer.  Even if multiple
Wrappers reference the same pointer, they all need to be dispose()'d
to make the ref count correct.