Making the dispose call back to synchronous.
The point of exposing the dispose call is to give applications a predictable release() behavior. If the dispose method only has the "we'll release it later" commitment, that's no different from not calling the method at all.

With this change NativePointerPhantomReference.releaseNative() can invoke twice, once by the dispose() call and the other when the Wrapper object actually gets GCed, but an earlier change I made in there makes this safe.