Implement support for SAFEARRAY return values
Add support for SAFEARRAY return values, SAFEARRAY in/out parameters,
and multi-dimensional SAFEARRAY's.

Notes:
This patch introduces fuller support for SAFEARRAYs.  This is
implemented by expanding the SafeArray class to represent a wrapper
over a native SAFEARRAY instance.  The managed-SafeArray owns the
native-SAFEARRAY, and will release it when it is garbage collected.
Conversions from single dimensional typed java arrays to SAFEARRAYS are
now primarily handled in managed code in the NativeType implementation.