Fixed issue within COM4J.loadNativeLibrary() where it will fail with a java.lang.ExceptionInInitializerError if the com4j JAR file is located within an UNC path.  Within that class, it will take the resource URL of COM4J.class (for example, jar:file://host/share/foo/lib/com4j-20120426-2.jar!/com4j/COM4J.class) to generate the path for the DLL to be extracted to.  In the example above, it will come up with the path of "\host\share\foo\lib\com4j-x86.dll", which is invalid.  This fixes the problem and returns a path starting with "\\host\share...".
Note:  The root cause of the ExceptionInInitializerError is actually:
Caused by: java.lang.NullPointerException
	at com4j.COM4J.loadNativeLibrary(COM4J.java:567)
	at com4j.COM4J.<clinit>(COM4J.java:522)
	... 7 more
It's actually not accurate, and it caused by the use of the LOGGER
static variable before it's initalized.  Though that only caused a less
meaningful stack trace, as the next step in the code should throw an
UnsatisfiedLinkError.  This is also fixed by moving the static LOGGER
initialization to right before the static block that calls
loadNativeLibrary().