Add -addLibNametoPackageName option to tlbimp
For some type libraries it can be useful to create seperate packages for
each library referenced by the type library.

For example, the Excel type library references Office and VBIDE and those
have conflicting class names if merged into a single package. Adding the
library name to the package name resolves that problem.